Fix %llu format specifier warnings
- Use %llu as format specifier
- Add D__USE_MINGW_ANSI_STDIO

+++ src/H5ls.c
@@ -75,7 +75,7 @@ herr_t H5Dget_info(hid_t d_id, H5D_info_t * dataset_info) {
 	int dims_char_written = 0;
 	int maxdims_char_written = 0;
 	for(int i = rank - 1; i >= 0; --i) {
-	  dims_char_written += sprintf(dataset_info->dims + dims_char_written, "%" PRIu64, (long long unsigned int) dims[i]);
+	  dims_char_written += sprintf(dataset_info->dims + dims_char_written, "%llu", dims[i]);
 	  if(i > 0) {
 	    dims_char_written += sprintf(dataset_info->dims + dims_char_written, " x ");
 	  }
@@ -85,7 +85,7 @@ herr_t H5Dget_info(hid_t d_id, H5D_info_t * dataset_info) {
 	    maxdims_char_written += sprintf(dataset_info->maxdims + maxdims_char_written, "Inf");
 	  }
 	  else {
-	    maxdims_char_written += sprintf(dataset_info->maxdims + maxdims_char_written, "%" PRIu64,  (long long unsigned int) maxdims[i]);
+	    maxdims_char_written += sprintf(dataset_info->maxdims + maxdims_char_written, "%llu", maxdims[i]);
 	  }
 	  if(i > 0) {
 	    maxdims_char_written += sprintf(dataset_info->maxdims + maxdims_char_written, " x ");
